cmd/install-wsl: capitalize all slog messages
Capitalize the first letter of all slog message strings (Info, Warn, Error)
for consistent formatting and improved readability. This treats each log
message as a complete sentence.

Also split main.go into domain-specific modules:
- github.go: GitHub API integration (fetchRelease, findMSIAsset)
- wsl.go: WSL operations (getInstalledWSLVersion, runInstallMSI, writeWslConfig)
- download.go: File download with atomic operations
- version.go: Version parsing and comparison
- main.go: Orchestration, httpClient, main(), run()

Tests remain in main_test.go for now; will be split in follow-up.

All tests pass.
